+++ Hermes_api/src/routes/adminFinance.js
@@ -1,8 +1,8 @@
-// File: src/routes/adminPayments.js
+// File: src/routes/adminFinance.js
 // Admin routes for manual payment confirmation
 const express = require('express');
 const router = express.Router();
-const adminPayments = require('../controllers/adminPayments');
+const adminFinance = require('../controllers/adminFinance');
 
 
 // Admin authentication middleware (requires HYPERADMIN role)
@@ -30,26 +30,26 @@ router.use(requireRole(['ADMIN', 'HYPERADMIN']));
 router.use(requireAdmin);
 
 // GET /admin/payments/pending - List pending mobile money deposits
-router.get('/pending', ...adminSecurity('admin-payments-pending'), adminPayments.getPendingPayments);
+router.get('/pending', ...adminSecurity('admin-payments-pending'), adminFinance.getPendingPayments);
 
 // POST /admin/payments/confirm - Manually confirm a payment and mint UGDX
-router.post('/confirm', ...adminSecurity('admin-payments-confirm'), adminPayments.confirmPayment);
+router.post('/confirm', ...adminSecurity('admin-payments-confirm'), adminFinance.confirmPayment);
 
 // POST /admin/payments/reject - Manually reject a payment
-router.post('/reject', ...adminSecurity('admin-payments-reject'), adminPayments.rejectPayment);
+router.post('/reject', ...adminSecurity('admin-payments-reject'), adminFinance.rejectPayment);
 
 // GET /admin/payments/history - Get payment history with filters
-router.get('/history', ...adminSecurity('admin-payments-history'), adminPayments.getPaymentHistory);
+router.get('/history', ...adminSecurity('admin-payments-history'), adminFinance.getPaymentHistory);
 
 // GET /admin/payments/treasury - Get onchain treasury overview and fee collections
 router.get('/treasury', 
   ...adminSecurity('admin-payments-treasury'), 
-  adminPayments.getTreasuryOverview
+  adminFinance.getTreasuryOverview
 );
 
 // GET /admin/payments/balance/:userId - Get user's on-chain UGDX balance
 router.get('/balance/:userId',
   ...adminSecurity('admin-payments-balance'),
-  adminPayments.getUserBalance
+  adminFinance.getUserBalance
 );
 module.exports = router;
